extract and split run_larc.py
